+++ src/Licensing/LicenseManager.php
@@ -144,6 +144,7 @@ class LicenseManager
             'testing' => $isTestDomain = $this->isOnTestDomain(),
             'message' => $this->invalidLicenseMessage($isTestDomain),
             'hasSiteKey' => $this->hasSiteKey(),
+            'sharedKey' => $this->site()->hasSharedKey(),
         ];
     }
 
@@ -196,6 +197,10 @@ class LicenseManager
 
     private function identityMessage(): string
     {
+        if ($this->site()->hasSharedKey()) {
+            return __('statamic::messages.licensing_shared_key');
+        }
+
         return $this->hasSiteKey()
             ? __('statamic::messages.licensing_site_key_found')
             : __('statamic::messages.licensing_site_key_missing');